+++ src/seqno_iter.rs
@@ -1,29 +1,22 @@
-use crate::{flatfile::FlatFile, Record, RecordSerializer};
-use std::sync::Arc;
+use crate::flatfile::FlatFile;
+use std::{mem::size_of, sync::Arc};
 
 /// This structure allows to iterate over records in the order they were added
 /// to this database. Note that this structure provides only the `Iterator`-like
 /// interface, because `Iterator` items cannot have associated lifetimes.
-pub struct SeqNoIter<R: RecordSerializer> {
+pub struct SeqNoIter {
     data: Arc<FlatFile>,
-    serializer: R,
     offset: usize,
 }
 
-impl<'a, R: RecordSerializer> SeqNoIter<R> {
-    pub(crate) fn new(data: Arc<FlatFile>, serializer: R, offset: usize) -> Self {
-        Self {
-            data,
-            serializer,
-            offset,
-        }
+impl SeqNoIter {
+    pub(crate) fn new(data: Arc<FlatFile>, offset: usize) -> Self {
+        Self { data, offset }
     }
 
-    pub fn next(&mut self) -> Option<Record> {
-        let item = self
-            .data
-            .get_record_at_offset(&self.serializer, self.offset)?;
-        self.offset += self.serializer.size(&item);
+    pub fn next(&mut self) -> Option<&[u8]> {
+        let item = self.data.get_record_at_offset(self.offset)?;
+        self.offset += item.len() + size_of::<u64>();
         Some(item)
     }
 }