+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Indexing by key - now records can only be indexed by their sequential number.
+  This also allows to remove serializers and `Record` type.
 
 ## [0.1.1] - 2019-07-22
 ### Fixed